prevent idiots (me) from killing their launchd (DO NOT USE UNTIL PRE-RELEASE)

+++ install_part1.sh
@@ -23,7 +23,9 @@ chmod 0755 /var/haxx
 chown mobile:mobile /var/haxx
 
 echo "copy files"
-cp /sbin/launchd /sbin/launchd.real
+if [ ! -f "/sbin/launchd.real" ]; then
+  cp /sbin/launchd /sbin/launchd.real
+fi
 cp ./files/haxx /usr/local/bin/haxx
 
 cp ./files/break.sh /usr/local/bin/break.sh
